Validator 분리
Number 과 String Validator 분리
