+++ src/main/java/calculator/StringValidator.java
@@ -7,15 +7,6 @@ public class StringValidator {
 
   private static final List<String> FOUR_OPERATOR = Arrays.asList("+", "-", "*", "/");
 
-  public static boolean isNumber(String val) {
-    try {
-      Integer.parseInt(val);
-      return true;
-    } catch (NumberFormatException nfe) {
-      return false;
-    }
-  }
-
   public static boolean isOperator(String val) {
     return FOUR_OPERATOR.contains(val);
   }
